Main

Documentation
Main' includes all functions to draw and build the game.
The most important function, main, has a PlayIO function running which triggers everything.
main :: IO ()

Display mode for the game; fullscreen
window :: Display

Has initial values for player, time and some scores in it. The rest of the initial game state is loaded from JSON.
initialState :: GTA

Key updates:
Notices all the key changes and triggers actions.
handleKeys :: Event -> GTA -> IO GTA

Sets keystates to state, in order to know if a key is being pressed (holded).
updateKeyState :: (KeyState, KeyState, KeyState, KeyState, Direction) -> GTA -> IO GTA

Pictures, rendering etc:
Base function from playIO to render the game (picture). Is mostly a list of other pictures, and defines the scal of the game + the focus around the player in the middle of the screen.
render :: GTA -> IO Picture

Shows the main (start) screen with information about the game.
mainScreen :: IO Picture

Function to show texts in the right scale and color.
text' :: (Float, Float) -> String -> Picture

Draws pictures that hold game state information.
statePicture :: GameState -> IO Picture

Holds the Picture with points (cash) and objectives in the top of the screen.
drawPoints :: GTA -> (Int, Int) -> Picture

Creates the timer in the top of the screen.
drawTimer :: GTA -> (Int, Int) -> Picture

Creates the text next to the timer.
timeLeftText :: GTA -> Picture

Game updates:
Base update function for game changes. The Running state triggers a few other update functions.
update :: Float -> GTA -> IO GTA

Checks if there's time left.
timeUp :: GTA -> GTA

Decides if a car can be entered or should be left.
enterOrLeaveCar :: GTA -> IO GTA

Varies gamestates, based on current gamestates.
changeGameState :: Char -> GTA -> IO GTA

State that's being used on startup to set everything in place.
loading :: GTA -> IO GTA

Generates a random int value, used in several decisions.
randomNr :: IO Int

Triggers all people and car updates (movements).
updateTraffic :: Int -> GTA -> GTA

Updates coins if a coin is reached.
updateCoins :: Float -> GTA -> GTA

Player, car and people, game(state) related updates
Triggers
updatePlayerPosition :: GTA -> GTA

Gives a subset of all blocks; all walkable/drivable for the player.

blocks' :: GTA -> [Block]

coins :: GTA -> [Block]

livingPeople :: GTA -> [Person]

killPerson :: GTA -> GTA

removeCoin :: GTA -> GTA

leaveCar :: GTA -> GTA

enterCar :: GTA -> GTA

updateCars :: [Car] -> Int -> GTA -> GTA

updateCar :: GTA -> Int -> Car -> (GameState, Car)

updatePeople :: [Person] -> Int -> GTA -> GTA

updatePerson :: GTA -> Int -> Person -> Person

close :: (Movable a, Movable b) => a -> [b] -> [Bool]