Models.Position

Documentation

Default data-type for a position.
data Position

Constructors

Position

Fields

x :: Float

y :: Float

Data-type for the Sprite, where spriteType links an object and spriteState to the state.
data Sprite

Constructors

Sprite

Fields

spriteType :: String

spriteState :: Int

Data-type for all possible directions.
data Direction

Constructors

North

East

South

West

Class Movable for Player, Person and Car to do something with the position and/or sprite.
class Movable a where

Minimal complete definition

getPos, setPos, getDir, setDir, width, height, getSprite

Methods

getPos :: a -> Position

setPos :: Position -> a -> a

getDir :: a -> Direction

setDir :: Direction -> a -> a

width :: a -> Float

height :: a -> Float

getSprite :: a -> Sprite

Draws the picture based on given arguments.
draw :: Movable a => [(String, Picture)] -> a -> Picture

Based on a given positions, moves an 'a'.
move :: Movable a => Position -> a -> a

Retrieves the coordinates of a given 'a'.
coordinates :: Movable a => a -> [(Float, Float)]

Gives the next direction, based on a given direction.
next :: Direction -> Direction

Gives the previous direction, based on a given direction.
prev :: Direction -> Direction

Changes the directions based on the given arguments.
changeDir :: Movable a => Int -> a -> a

Self defined function to round numbers.
roundDecimals :: (Fractional a2, RealFrac a1, Integral b) => a1 -> b -> a2

Gives the next sprite state, based on a given sprite.
nextSprite :: Sprite -> Int

Changes the direction randomly.
changeDirR :: Movable a => Int -> a -> a